update styling on job modal and separate into separate components
