+++ src/App.tsx
@@ -14,10 +14,10 @@ import { motion, AnimatePresence } from "framer-motion";
 import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
 import { Button } from "@/components/ui/button";
 import { Header } from "@/components/ui/header";
-import { QueueStatus } from "@/components/ui/queue-status";
-import { QueueGrid } from "@/components/ui/queue-grid";
-import { WorkflowDiagram } from "@/components/ui/workflow-diagram";
-import { DebugView } from "@/components/ui/debug-view";
+import { QueueStatus } from "@/components/queue-status";
+import { QueueGrid } from "@/components/queue-grid";
+import { WorkflowDiagram } from "@/components/workflow-diagram";
+import { DebugView } from "@/components/debug-view";
 import { Toaster } from "@/components/ui/sonner";
 import { toast } from "sonner";
 import { useQueues } from "@/hooks/useQueues";
