+++ src/components/debug-view.tsx
