+++ src/components/ui/header.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { motion } from "framer-motion";
 import { FileText, Bot } from "lucide-react";
