+++ src/components/wikidata-preview.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { ExternalLink } from "lucide-react";
 
 interface WikidataProps {
